Fixed scrolling with more than one mutation record
I have a logic, where styles of the message bubbles are formatted like on
facebook - first bubble is rounded everywhere, than next bubble causes
first to change it's corner roundness, etc.
Because of that, there were multiple MutationRecord objects in "e"
object and only the first was the added node. Now your code checks
the last MutationRecord, and it was only the removed node.
I don't know if it will be universal to all cases, but maybe consider
adjustable index of the "e" array?