feat: add ability to handle item prepending
When the handlePrepend config flag is set to true, v-chat-scroll will
keep track of the changing scrollHeight of the container. When scrollTop
is zero (ie. the user's scrolled all the way to the top) and items are
added, v-chat-scroll will scroll to keep the previously viewed item in
place by scrolling to the difference in scrollHeight before and after
the new items were added.